+++ src/ConsoleL/Interface/CommandInterface.php
@@ -20,4 +20,6 @@ interface CommandInterface
 		array $arguments,
 		string $baseDir
 	): void;
+
+	public static function generateSecretKey(array $arguments): void;
 }