+++ src/Foundation/Render.php
@@ -18,50 +18,56 @@ final class Render extends Controller
 		self::$request_uri = Application::$request_uri;
 	}
 
+	/**
+	 * Render the web route - primary routing
+	 */
 	public static function WebRoute()
 	{
 		self::Load();
+		$static = new static();
 		$reg_route = $GLOBALS['__registered_routes'];
 
 		foreach ($reg_route as $route) {
+			self::$middleware = $route['middleware'];
+			self::$map_info = $route['map_info'];
+			self::$redirect = $route['redirect'];
+			self::$method = $route['method'];
+			self::$action = $route['action'];
+			self::$view = $route['view'];
+			self::$file = $route['file'];
+			self::$any = $route['any'];
+			self::$use = $route['use'];
+
 			if ($route['middleware']) {
-				self::$middleware = $route['middleware'];
-				self::__middleware();
+				$static->__middleware();
 			}
 
 			if ($route['redirect']) {
-				self::$redirect = $route['redirect'];
 				self::__redirect();
 			}
 
-			if ($route['action']) {
-				self::$action = $route['action'];
-				self::__action();
-			}
-
-			if ($route['any']) {
-				self::$any = $route['any'];
-				self::__any();
-			}
-
-			if ($route['use']) {
-				self::$use = $route['use'];
-				self::__use();
-			}
-
-			if ($route['file']) {
-				self::$file = $route['file'];
-				self::__file();
-			}
-
 			if ($route['method']) {
-				self::$method = $route['method'];
 				self::__method();
 			}
 
 			if ($route['view']) {
-				self::$view = $route['view'];
 				self::__view();
+			}
+
+			if ($route['any']) {
+				self::__any();
+			}
+
+			if ($route['use']) {
+				$static->__use();
+			}
+
+			if ($route['file']) {
+				$static->__file();
+			}
+
+			if ($route['action']) {
+				$static->__action();
 			}
 		}
 	}