+++ test/Web/JwtTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use PhpSlides\Services\JwtService;
+use PhpSlides\Web\JWT;
 
 include_once __DIR__ . '/../autoload.php';
 include_once __DIR__ . '/../../src/Globals/Functions.php';
@@ -13,17 +13,17 @@ $payload = payload(data: ['user_id' => '555'], expires: time() + 3600);
 /**
  * Testing JwtService encode method
  */
-$token = JwtService::encode($payload);
+$token = JWT::encode($payload);
 
 /**
  * Testing JwtService verify token method
  */
-$verifyToken = JwtService::verify($token);
+$verifyToken = JWT::verify($token);
 
 if ($verifyToken) {
 	/**
 	 * Testing JwtService decode method
 	 */
-	$decodedToken = JwtService::decode($token);
+	$decodedToken = JWT::decode($token);
 	print_r($decodedToken);
 }